+++ tests/testthat/test-describe_distribution.R
@@ -242,6 +242,31 @@ test_that("describe_distribution - grouped df", {
   expect_equal(out$Mean, c(1.462, 0.246, 4.26, 1.326, 5.552, 2.026), tolerance = 1e-3)
 })
 
+test_that("argument 'by' works", {
+  skip_if_not_installed("bayestestR")
+
+  # basic
+  grouped <- data_group(mtcars, c("am", "vs"))
+  expect_identical(
+    describe_distribution(grouped),
+    describe_distribution(mtcars, by = c("am", "vs")),
+    ignore_attr = TRUE
+  )
+
+  # mixing data_group() and arg 'by'
+  grouped <- data_group(mtcars, c("am", "vs"))
+  half_grouped <- data_group(mtcars, "am")
+  expect_identical(
+    describe_distribution(grouped),
+    describe_distribution(half_grouped, by = "vs"),
+    ignore_attr = TRUE
+  )
+
+  expect_error(
+    describe_distribution(mtcars, by = 2),
+    "must be a character vector"
+  )
+})
 
 # distribution_mode --------------------------
 test_that("distribution_mode works as expected", {